refactor: :truck: Move data api and shared prefs files
